net, kcov, mac80211: collect coverage from specific socket buffers
This patch enables Linux to remember the specific kcov handle of a
thread that allocated/copied a specific socket buffer.

This information is then passed to kcov_remote_start()/stop()
annotations in order to collect kcov coverage for the code that
processes incoming 802.11 frames. Later this stored kcov_handle can be
used to enable kcov coverage collection for other network operations.