kernel/kcov.c: properly handle subsequent mmap calls
Subsequent mmaps of the same kcov descriptor actually do not update the
virtual memory of the task and yet return 0 (success). This is
counter-intuitive and may lead to unexpected SIGBUS errors.

Also, this unnecessarily limits the functionality of kcov to only the
simplest usage scenarios. E.g. due to this problem syzkaller cannot
reliably delegate kcov mmapping to its forked children. Doing so
essentially prevents the reusage of kcov instances - they get forever
attached to their first address space.

Refactor the kcov_mmap function so that it always inserts pages into the
virtual memory of the calling process, but allocates the memory only
during the first call.

Signed-off-by: Aleksandr Nogikh <nogikh@google.com>